// ui/src/app/core/api-client/model/server-instance-dto.ts
/**
 * API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ServerInstanceDto { 
    name?: string;
    type?: ServerInstanceDto.TypeEnum;
    cpu?: number;
    memory?: number;
    eula?: boolean;
    version?: string;
    modrinthProjectId?: string;
    curseforgePageUrl?: string;
}
export namespace ServerInstanceDto {
    export const TypeEnum = {
        Vanilla: 'VANILLA',
        Plugin: 'PLUGIN',
        Mod: 'MOD',
        Modrinth: 'MODRINTH',
        Curseforge: 'CURSEFORGE'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}